Move product_dict to utils.py
